Create layoutlib artifacts for GMaven
This creates a gradle project for building artifacts for layoutlib to be
shipped on GMaven. One artifact is simply the layoutlib jar. Another one
is composed of all the native libraries and associated data files
(fonts, icu...). This latter artifact, named runtime, is using Gradle
module metadata to only ship the relevant files for a given
architecture/OS.

Bug: 241795063
Test: N/A
Change-Id: I047218ec29a95bd5ec7def4cd30b4544aac47e30